OCT-111: moved GetChannelsQuery
