+++ components/catalogs/back/tests/Integration/Infrastructure/Persistence/Channel/GetChannelsQueryTest.php
@@ -2,16 +2,16 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Catalogs\Test\Integration\Infrastructure\Persistence;
+namespace Akeneo\Catalogs\Test\Integration\Infrastructure\Persistence\Channel;
 
-use Akeneo\Catalogs\Infrastructure\Persistence\GetChannelsQuery;
+use Akeneo\Catalogs\Infrastructure\Persistence\Channel\GetChannelsQuery;
 use Akeneo\Catalogs\Test\Integration\IntegrationTestCase;
 
 /**
  * @copyright 2022 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  *
- * @covers \Akeneo\Catalogs\Infrastructure\Persistence\GetChannelsQuery
+ * @covers \Akeneo\Catalogs\Infrastructure\Persistence\Channel\GetChannelsQuery
  */
 class GetChannelsQueryTest extends IntegrationTestCase
 {
