+++ components/catalogs/back/tests/Integration/Infrastructure/Persistence/GetChannelsByCodeQueryTest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Akeneo\Catalogs\Test\Integration\Infrastructure\Persistence;
 
-use Akeneo\Catalogs\Infrastructure\Persistence\GetChannelsByCodeQuery;
+use Akeneo\Catalogs\Infrastructure\Persistence\Channel\GetChannelsByCodeQuery;
 use Akeneo\Catalogs\Test\Integration\IntegrationTestCase;
 
 /**
@@ -12,7 +12,7 @@ use Akeneo\Catalogs\Test\Integration\IntegrationTestCase;
  * @copyright 2022 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  *
- * @covers \Akeneo\Catalogs\Infrastructure\Persistence\GetChannelsByCodeQuery
+ * @covers \Akeneo\Catalogs\Infrastructure\Persistence\Channel\GetChannelsByCodeQuery
  */
 class GetChannelsByCodeQueryTest extends IntegrationTestCase
 {
